refactor: GameStartPage/ArrowControl
