+++ packages/client/src/pages/GameStart/components/ArrowControl/components/Arrow/index.tsx
