--- packages/client/src/pages/GameStart/components/ArrowControl/index.ts
@@ -1,1 +1,0 @@
-export { ArrowControl } from './ArrowControl'
